+++ src/components/Category/CategoryForm.tsx
@@ -1,10 +1,5 @@
 import React, { useState, useEffect, ChangeEvent, FormEvent } from 'react';
 
-interface CategoryFormProps {
-  onSubmit: (categoryName: string) => void;
-  initialCategory?: string;
-}
-
 const CategoryForm: React.FC<CategoryFormProps> = ({
   onSubmit,
   initialCategory = '',
