+++ src/components/auth/Form.tsx
@@ -2,11 +2,6 @@ import React from 'react';
 import { FC, useState } from 'react';
 import './Form.css';
 
-interface FormProps {
-  title: string;
-  handleClick: (email: string, pass: string) => void;
-}
-
 const Form: FC<FormProps> = ({ title, handleClick }) => {
   const [email, setEmail] = useState('');
   const [pass, setPass] = useState('');
