+++ src/index.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 import App from './App';
-import { store, persistor } from './store'; // Убедитесь, что persistor экспортируется из файла store
+import { store, persistor } from './store';
 import './firebase/firebase';
 import './global.css';
